Restructure agent package and update tests
