+++ agent/__init__.py
@@ -1,9 +1,9 @@
-from .base_agent import BaseAgent
-from .coder_agent import CoderAgent
-from .config import NiraConfig, load_config
-from .nira_agent import NiraAgent
-from .researcher_agent import ResearcherAgent
-from .sysops_agent import SysOpsAgent
+from .agents.base_agent import BaseAgent
+from .agents.coder_agent import CoderAgent
+from .core.config import NiraConfig, load_config
+from .agents.nira_agent import NiraAgent
+from .agents.researcher_agent import ResearcherAgent
+from .agents.sysops_agent import SysOpsAgent
 
 __all__ = [
     "BaseAgent",
