+++ agent/agents/coder_agent.py
@@ -1,6 +1,6 @@
 from .base_agent import BaseAgent
-from .tools.github_manager_tool import github_manager
-from .tools.run_bash_command_tool import run_bash_command_tool
+from ..tools.coder.github_manager_tool import github_manager
+from ..tools.sysops.run_bash_command_tool import run_bash_command_tool
 
 CODER_SYSTEM = "You are a coding assistant helping with programming tasks."
 CODER_TOOLS = [run_bash_command_tool, github_manager]
