+++ agent/agents/researcher_agent.py
@@ -1,7 +1,7 @@
 from .base_agent import BaseAgent
-from .tools.summarise_text_tool import summarise_text_tool
-from .tools.summarize_youtube_tool import summarize_youtube_tool
-from .tools.web_search_tool import web_search_tool
+from ..tools.researcher.summarise_text_tool import summarise_text_tool
+from ..tools.researcher.summarize_youtube_tool import summarize_youtube_tool
+from ..tools.researcher.web_search_tool import web_search_tool
 
 RESEARCHER_SYSTEM = "You research information and summarise findings."
 RESEARCHER_TOOLS = [web_search_tool, summarise_text_tool, summarize_youtube_tool]
