+++ agent/agents/sysops_agent.py
@@ -1,6 +1,6 @@
 from .base_agent import BaseAgent
-from .tools.check_website_tool import check_website_tool
-from .tools.run_bash_command_tool import run_bash_command_tool
+from ..tools.sysops.check_website_tool import check_website_tool
+from ..tools.sysops.run_bash_command_tool import run_bash_command_tool
 
 SYSOPS_SYSTEM = "You perform system administration and operational tasks."
 SYSOPS_TOOLS = [run_bash_command_tool, check_website_tool]
