+++ agent/core/__init__.py
@@ -1,0 +1,12 @@
+from .config import NiraConfig, load_config
+from .status import console, status_manager
+from .metrics import init_metrics, track_tool
+
+__all__ = [
+    "NiraConfig",
+    "load_config",
+    "console",
+    "status_manager",
+    "init_metrics",
+    "track_tool",
+]
