+++ agent/tools/coder/github_manager_tool.py
@@ -1,9 +1,9 @@
 from langchain_core.tools import tool
 from pydantic import BaseModel, Field
 
-from ..config import load_config
-from ..metrics import track_tool
-from .http_utils import request_json
+from ...core.config import load_config
+from ...core.metrics import track_tool
+from ..http_utils import request_json
 
 
 class GitHubManagerInput(BaseModel):
