+++ agent/tools/file_manager_tool.py
@@ -4,8 +4,8 @@ from typing import List
 from langchain_core.tools import tool
 from pydantic import BaseModel, Field, NonNegativeInt
 
-from ..metrics import track_tool
-from ..status import status_manager
+from ..core.metrics import track_tool
+from ..core.status import status_manager
 
 
 class FileManagerInput(BaseModel):
