+++ agent/tools/get_domain_info_tool.py
@@ -2,8 +2,8 @@ import dns.resolver
 from langchain_core.tools import tool
 from pydantic import BaseModel, Field
 
-from ..config import load_config
-from ..metrics import track_tool
+from ..core.config import load_config
+from ..core.metrics import track_tool
 
 
 class DomainInfoInput(BaseModel):
