+++ agent/tools/sysops/check_website_tool.py
@@ -2,7 +2,7 @@ import requests
 from langchain_core.tools import tool
 from pydantic import BaseModel, Field
 
-from ..metrics import track_tool
+from ...core.metrics import track_tool
 
 
 class CheckWebsiteInput(BaseModel):
