+++ tests/test_agent.py
@@ -2,7 +2,7 @@ import unittest
 
 from langchain_community.llms import FakeListLLM
 
-from agent.base_agent import BaseAgent
+from agent.agents.base_agent import BaseAgent
 
 
 class AgentConfigTest(unittest.TestCase):
