+++ tests/test_github_tool.py
@@ -1,11 +1,11 @@
 import unittest
 from unittest.mock import patch
 
-from agent.tools.github_manager_tool import github_manager
+from agent.tools.coder.github_manager_tool import github_manager
 
 
 class GitHubToolTest(unittest.TestCase):
-    @patch("agent.tools.github_manager_tool.request_json")
+    @patch("agent.tools.coder.github_manager_tool.request_json")
     def test_get_repo_info_success(self, mock_get):
         mock_get.return_value = {
             "full_name": "octocat/Hello-World",
@@ -18,14 +18,14 @@ class GitHubToolTest(unittest.TestCase):
         mock_get.assert_called_once()
 
     @patch(
-        "agent.tools.github_manager_tool.request_json",
+        "agent.tools.coder.github_manager_tool.request_json",
         return_value="Failed to fetch repo info: fail",
     )
     def test_get_repo_info_error(self, mock_get):
         result = github_manager.func("repo_info", repo="octocat/Hello-World")
         self.assertIn("Failed to fetch", result)
 
-    @patch("agent.tools.github_manager_tool.request_json")
+    @patch("agent.tools.coder.github_manager_tool.request_json")
     def test_create_repo_success(self, mock_post):
         mock_post.return_value = {"name": "new"}
         with patch.dict("os.environ", {"GITHUB_TOKEN": "t"}):
@@ -35,7 +35,7 @@ class GitHubToolTest(unittest.TestCase):
         mock_post.assert_called_once()
 
     @patch(
-        "agent.tools.github_manager_tool.request_json",
+        "agent.tools.coder.github_manager_tool.request_json",
         return_value="Failed to create repo: fail",
     )
     def test_create_repo_error(self, mock_post):
@@ -43,7 +43,7 @@ class GitHubToolTest(unittest.TestCase):
             result = github_manager.func("create_repo", repo="new")
         self.assertIn("Failed to create repo", result)
 
-    @patch("agent.tools.github_manager_tool.request_json")
+    @patch("agent.tools.coder.github_manager_tool.request_json")
     def test_create_issue_success(self, mock_post):
         mock_post.return_value = {"title": "bug"}
         with patch.dict("os.environ", {"GITHUB_TOKEN": "t"}):
@@ -53,7 +53,7 @@ class GitHubToolTest(unittest.TestCase):
         self.assertEqual(result["title"], "bug")
         mock_post.assert_called_once()
 
-    @patch("agent.tools.github_manager_tool.request_json")
+    @patch("agent.tools.coder.github_manager_tool.request_json")
     def test_create_pr_success(self, mock_post):
         mock_post.return_value = {"title": "pr"}
         with patch.dict("os.environ", {"GITHUB_TOKEN": "t"}):
