+++ tests/test_planner_executor.py
@@ -2,9 +2,9 @@ import unittest
 
 from langchain_community.llms import FakeListLLM
 
-from agent.coder_agent import CoderAgent
-from agent.nira_agent import NiraAgent
-from agent.planner_executor import PlannerExecutor
+from agent.agents.coder_agent import CoderAgent
+from agent.agents.nira_agent import NiraAgent
+from agent.agents.planner_executor import PlannerExecutor
 
 
 class PlannerExecutorTest(unittest.TestCase):
